refactor: Move start/stop OrbitDB convenience functions to own repo.
