+++ src/utils/index.js
@@ -1,6 +1,5 @@
 import createId from './create-id.js'
-import startOrbitDB from './start.js'
+
 export {
-  createId,
-  startOrbitDB
+  createId
 }
