+++ test/utils/create-helia.js
@@ -3,12 +3,70 @@ import { bitswap } from '@helia/block-brokers'
 import { createLibp2p } from 'libp2p'
 import { MemoryBlockstore } from 'blockstore-core'
 import { LevelBlockstore } from 'blockstore-level'
-import { DefaultLibp2pOptions, DefaultLibp2pBrowserOptions } from '../../src/index.js'
+import { identify } from '@libp2p/identify'
+import { webSockets } from '@libp2p/websockets'
+import { webRTC } from '@libp2p/webrtc'
+import { all } from '@libp2p/websockets/filters'
+import { noise } from '@chainsafe/libp2p-noise'
+import { yamux } from '@chainsafe/libp2p-yamux'
+import { gossipsub } from '@chainsafe/libp2p-gossipsub'
+import { circuitRelayTransport } from '@libp2p/circuit-relay-v2'
 
 const isBrowser = () => typeof window !== 'undefined'
 
+const Libp2pOptions = {
+  addresses: {
+    listen: ['/ip4/0.0.0.0/tcp/0/ws']
+  },
+  transports: [
+    webSockets({
+      filter: all
+    }),
+    webRTC(),
+    circuitRelayTransport({
+      discoverRelays: 1
+    })
+  ],
+  connectionEncryption: [noise()],
+  streamMuxers: [yamux()],
+  connectionGater: {
+    denyDialMultiaddr: () => false
+  },
+  services: {
+    identify: identify(),
+    pubsub: gossipsub({ allowPublishToZeroPeers: true })
+  }
+}
+
+/**
+ * A basic Libp2p configuration for browser nodes.
+ */
+const Libp2pBrowserOptions = {
+  addresses: {
+    listen: ['/webrtc']
+  },
+  transports: [
+    webSockets({
+      filter: all
+    }),
+    webRTC(),
+    circuitRelayTransport({
+      discoverRelays: 1
+    })
+  ],
+  connectionEncryption: [noise()],
+  streamMuxers: [yamux()],
+  connectionGater: {
+    denyDialMultiaddr: () => false
+  },
+  services: {
+    identify: identify(),
+    pubsub: gossipsub({ allowPublishToZeroPeers: true })
+  }
+}
+
 export default async ({ directory } = {}) => {
-  const options = isBrowser() ? DefaultLibp2pBrowserOptions : DefaultLibp2pOptions
+  const options = isBrowser() ? Libp2pBrowserOptions : Libp2pOptions
 
   const libp2p = await createLibp2p({ ...options })
